Refacto: Separate signatures and modules, allow to save outputs and inputs
